requirements.txt
asgiref==3.8.1
babel==2.17.0
diff-match-patch==20241021
Django==5.2.1
django-autoslug==1.9.9
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-grappelli==4.0.2
django-import-export==4.3.7
django-money==3.5.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
Faker==37.3.0
pillow==11.2.1
py-moneyed==3.0
PyJWT==2.9.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
